[API] Correct /api routing for new backend
