+++ api/src/routes/index.js
@@ -1,64 +1,14 @@
-const config = require('../config')
 const router = require('express').Router();
 const auth = require('./authentication')
-const events = require('../lib/eventlist')
+const base = require('./base')
 
 /**
- * Non-authenticated Routes
+ * Base Routes
  */
-
-/**
- * GET /
- */
-router.get('/', (req, res) => res.status(200));
-
-/**
- * GET config
- */
-router.get('/config',
-  function (req, res) {
-    res.status(200).send({
-      SiteHost: config.SiteHost,
-      APIUrl: config.APIUrl,
-      APIAuthBasePath: config.APIUrl + config.APIAuthBasePath,
-      APICurrentEventUrl: config.APIUrl + config.APIEventsUrl + '/current',
-      APINextEventUrl: config.APIUrl + config.APIEventsUrl + '/next',
-      APILoginUrl: config.APIUrl + config.APILoginUrl,
-      APIAuthUrl: config.APIUrl + config.APIAuthUrl,
-      ports: config.ports,
-      soundcloud: config.soundcloud
-    })
-  }
-)
-
-/**
- * GET Events - Current
- */
-router.get('/event/current', async (req, res) => {
-  try {
-    let ev = await events.getEvents(true)
-    res.status(200).send(ev.current)
-  } catch (err) {
-    res.status(500, { error: err })
-  }
-})
-
-
-/**
- * GET Events - Next
- */
-router.get('/event/next', async (req, res) => {
-  try {
-    let ev = await events.getEvents(true)
-    res.status(200).send(ev.next)
-  } catch (err) {
-    res.status(500, { error: err })
-  }
-})
-
+router.use('/api', base);
 /**
  * Authenticated Routes
  */
-router.use('/auth', auth);
+router.use('/api/auth', auth);
 
 module.exports = router;
